# openapi.json
{
  "components": {
    "schemas": {
      "ClusterContainer": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "pid_scope": {
            "$ref": "#/components/schemas/PidScope"
          }
        },
        "required": [
          "id",
          "name",
          "image",
          "pid",
          "pid_scope"
        ],
        "title": "ClusterContainer",
        "type": "object"
      },
      "ClusterEndpoint": {
        "additionalProperties": false,
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          }
        },
        "required": [
          "host",
          "port"
        ],
        "title": "ClusterEndpoint",
        "type": "object"
      },
      "ClusterMetrics": {
        "additionalProperties": false,
        "properties": {
          "cpu_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "memory_limit_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "memory_usage_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "sampled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "volume_usage_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "cpu_percent",
          "memory_usage_bytes",
          "memory_limit_bytes",
          "volume_usage_bytes",
          "sampled_at"
        ],
        "title": "ClusterMetrics",
        "type": "object"
      },
      "ClusterSnapshot": {
        "additionalProperties": false,
        "properties": {
          "container": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ClusterContainer"
              }
            ]
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ClusterEndpoint"
              }
            ]
          },
          "metrics": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ClusterMetrics"
              }
            ]
          },
          "mode": {
            "enum": [
              "compose",
              "external"
            ]
          },
          "owned": {
            "type": "boolean"
          },
          "sampled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/PostgresClusterState"
          },
          "unavailability_reason": {
            "anyOf": [
              {
                "enum": [
                  "docker_unavailable",
                  "external_not_owned",
                  "inspect_failed",
                  "missing",
                  "stats_failed",
                  "stopped"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "mode",
          "owned",
          "state",
          "endpoint",
          "container",
          "metrics",
          "unavailability_reason",
          "sampled_at"
        ],
        "title": "ClusterSnapshot",
        "type": "object"
      },
      "DatabaseFootprint": {
        "additionalProperties": false,
        "properties": {
          "filestore_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "owned": {
            "type": "boolean"
          },
          "postgres_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "owned",
          "postgres_bytes",
          "filestore_bytes",
          "total_bytes"
        ],
        "title": "DatabaseFootprint",
        "type": "object"
      },
      "EnvironmentArtifacts": {
        "additionalProperties": false,
        "properties": {
          "backup_exists": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ]
          },
          "config_exists": {
            "type": "boolean"
          },
          "dependency_lock_exists": {
            "type": "boolean"
          },
          "python_contained": {
            "type": "boolean"
          },
          "python_exists": {
            "type": "boolean"
          },
          "worktree_exists": {
            "type": "boolean"
          },
          "worktree_registered": {
            "type": "boolean"
          }
        },
        "required": [
          "worktree_exists",
          "worktree_registered",
          "config_exists",
          "python_exists",
          "python_contained",
          "dependency_lock_exists",
          "backup_exists"
        ],
        "title": "EnvironmentArtifacts",
        "type": "object"
      },
      "EnvironmentSnapshot": {
        "additionalProperties": false,
        "properties": {
          "allocated_http_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifacts": {
            "$ref": "#/components/schemas/EnvironmentArtifacts"
          },
          "branch": {
            "type": "string"
          },
          "database": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "db_mode": {
            "enum": [
              "copy",
              "shared"
            ]
          },
          "git": {
            "$ref": "#/components/schemas/GitActivity"
          },
          "id": {
            "type": "string"
          },
          "lifecycle_state": {
            "$ref": "#/components/schemas/EnvironmentState"
          },
          "name": {
            "type": "string"
          },
          "observed_port": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PortObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "pgadmin": {
            "$ref": "#/components/schemas/PgAdminEligibility"
          },
          "project_id": {
            "type": "string"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeMetrics"
          },
          "short_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "storage": {
            "$ref": "#/components/schemas/StorageFootprint"
          }
        },
        "required": [
          "id",
          "project_id",
          "name",
          "branch",
          "short_sha",
          "db_mode",
          "database",
          "lifecycle_state",
          "allocated_http_port",
          "observed_port",
          "artifacts",
          "runtime",
          "git",
          "storage",
          "pgadmin"
        ],
        "title": "EnvironmentSnapshot",
        "type": "object"
      },
      "EnvironmentState": {
        "description": "Persisted lifecycle state shared by catalog and monitor contracts.",
        "enum": [
          "cleanup_failed",
          "creating",
          "failed",
          "ready",
          "removed",
          "removing"
        ],
        "title": "EnvironmentState"
      },
      "GitActivity": {
        "additionalProperties": false,
        "properties": {
          "ahead": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "behind": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "branch": {
            "type": "string"
          },
          "default_branch": {
            "type": "string"
          },
          "diff": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GitDiff"
              }
            ]
          },
          "head_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "short_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/GitActivityState"
          }
        },
        "required": [
          "default_branch",
          "head_sha",
          "short_sha",
          "branch",
          "ahead",
          "behind",
          "diff",
          "state"
        ],
        "title": "GitActivity",
        "type": "object"
      },
      "GitActivityState": {
        "enum": [
          "ahead",
          "behind",
          "clean",
          "diverged",
          "orphan"
        ],
        "title": "GitActivityState"
      },
      "GitDiff": {
        "additionalProperties": false,
        "properties": {
          "added": {
            "type": "integer"
          },
          "deleted": {
            "type": "integer"
          }
        },
        "required": [
          "added",
          "deleted"
        ],
        "title": "GitDiff",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HttpError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/HttpErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HttpError",
        "type": "object"
      },
      "HttpErrorCode": {
        "enum": [
          "database_not_found",
          "environment_not_found",
          "invalid_request",
          "monitor_snapshot_failed",
          "pgadmin_not_eligible",
          "pgadmin_unavailable"
        ],
        "title": "HttpErrorCode"
      },
      "PgAdminEligibility": {
        "additionalProperties": false,
        "properties": {
          "state": {
            "$ref": "#/components/schemas/PgAdminEligibilityState"
          }
        },
        "required": [
          "state"
        ],
        "title": "PgAdminEligibility",
        "type": "object"
      },
      "PgAdminEligibilityState": {
        "enum": [
          "cluster_not_owned",
          "cluster_unhealthy",
          "database_unresolved",
          "eligible",
          "environment_not_ready"
        ],
        "title": "PgAdminEligibilityState"
      },
      "PgAdminOpenRequest": {
        "additionalProperties": false,
        "properties": {
          "environment_id": {
            "type": "string"
          }
        },
        "required": [
          "environment_id"
        ],
        "title": "PgAdminOpenRequest",
        "type": "object"
      },
      "PgAdminOpenResult": {
        "additionalProperties": false,
        "properties": {
          "state": {
            "$ref": "#/components/schemas/PgAdminOpenState"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "state",
          "url"
        ],
        "title": "PgAdminOpenResult",
        "type": "object"
      },
      "PgAdminOpenState": {
        "enum": [
          "reconfigured",
          "reused",
          "started"
        ],
        "title": "PgAdminOpenState"
      },
      "PidScope": {
        "enum": [
          "docker_vm",
          "host",
          "unavailable"
        ],
        "title": "PidScope"
      },
      "PortObservation": {
        "enum": [
          "free",
          "occupied",
          "unknown"
        ],
        "title": "PortObservation"
      },
...
      "ProjectSummary": {
        "additionalProperties": false,
        "properties": {
          "cluster": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ClusterSnapshot"
              }
            ]
          },
          "display_hint": {
            "type": "string"
          },
          "environment_count": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "display_hint",
          "environment_count",
          "cluster"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
      "PythonEnvFootprint": {
        "additionalProperties": false,
        "properties": {
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "owned": {
            "type": "boolean"
          }
        },
        "required": [
          "owned",
          "bytes"
        ],
        "title": "PythonEnvFootprint",
        "type": "object"
      },
      "RuntimeMetrics": {
        "additionalProperties": false,
        "properties": {
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "child_pids": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "cpu_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "http_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "http_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "process_count": {
            "type": "integer"
          },
          "root_pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "rss_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/RuntimeState"
          }
        },
        "required": [
          "state",
          "root_pid",
          "child_pids",
          "process_count",
          "cpu_percent",
          "rss_bytes",
          "started_at",
          "http_url",
          "http_port",
          "database_name",
          "commit_sha",
          "branch"
        ],
        "title": "RuntimeMetrics",
        "type": "object"
      },
      "RuntimeState": {
        "enum": [
          "not_ready",
          "ready",
          "stopped"
        ],
        "title": "RuntimeState"
      },
      "Snapshot": {
        "additionalProperties": false,
        "properties": {
          "environments": {
            "items": {
              "$ref": "#/components/schemas/EnvironmentSnapshot"
            },
            "type": "array"
          },
          "generated_at": {
            "type": "string"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "type": "array"
          },
          "schema_version": {
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "generated_at",
          "projects",
          "environments"
        ],
        "title": "Snapshot",
        "type": "object"
      },
      "StorageFootprint": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "type": "boolean"
          },
          "database": {
            "$ref": "#/components/schemas/DatabaseFootprint"
          },
          "other_files_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "python_environment": {
            "$ref": "#/components/schemas/PythonEnvFootprint"
          },
          "total_bytes": {
            "type": "integer"
          },
          "worktree_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "total_bytes",
          "complete",
          "worktree_bytes",
          "python_environment",
          "database",
          "other_files_bytes"
        ],
        "title": "StorageFootprint",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "odoo-instance-sdk monitor",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/pgadmin/open": {
      "post": {
        "operationId": "openPgAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PgAdminOpenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgAdminOpenResult"
                }
              }
            },
            "description": "pgAdmin opened"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            },
            "description": "Environment not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            },
            "description": "pgAdmin operation rejected"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            },
            "description": "Invalid request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            },
            "description": "pgAdmin unavailable"
          }
        },
        "summary": "Open Pgadmin"
      }
    },
    "/api/v1/snapshot": {
      "get": {
        "operationId": "getMonitorSnapshot",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            },
            "description": "Monitor snapshot failed"
          }
        },
        "summary": "Snapshot"
      }
    },
...
  }
}